requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
pydantic
dynaconf>=3.2.0
httpx==0.25.2
aiohttp==3.9.1
jsonschema==4.20.0
pyyaml==6.0.1
sqlalchemy==2.0.35
alembic==1.13.1
aiosqlite==0.19.0
greenlet
plotly==6.4.0